refactor(main): extract system-modal popup into popups/systemModal
Amp-Thread-ID: https://ampcode.com/threads/T-019e1fcf-7754-703e-9441-11cc2816c591
